Adds zookeeper cluster support
Zookeeper will be assembled into a cluster,
which is pretty important for production environments.

Changes were tested in Ubuntu containers.

It seems, that CentOS support was initially broken for some time,
because zookeeper package is absent in CentOS 7.

Added handler to restart zookeeper on config change.

Might be used as openstack ceilometer/gnocchi coordinator
or for other apps using Tooz project